+++ src/Textarea/table.tsx
@@ -1,5 +1,5 @@
 import { ModalProvider } from 'styled-react-modal'
-import { StyledModal } from './styled'
+import { StyledModal } from '../styled/styled'
 import React, { ChangeEvent, useState } from 'react'
