+++ src/main.css
@@ -1,192 +1,187 @@
-
 .full-height {
-    height: 100%;
+  height: 100%;
 }
 
 /* 标题 */
 .header {
-    padding: 1rem 1rem;
-    background-color: rgb(187, 147, 1);
-    font-size: 40px;
+  padding: 1rem 1rem;
+  background-color: rgb(187, 147, 1);
+  font-size: 40px;
 }
 
-.header i{
-    padding: 1rem;
-    font-size: 48px;
+.header i {
+  padding: 1rem;
+  font-size: 48px;
 }
 
-.themeButton{
-    border:none;
-    background-color: transparent;
+.themeButton {
+  border: none;
+  background-color: transparent;
 }
 
 /*工具栏区*/
 
 .toolbars {
-    background-color: beige;
-    border: solid rgb(178, 178, 104);
-    border-radius: 10px;
-    display: flex;
-    justify-content: center;
-    margin: 1rem 24%;
-    align-items: center;
+  background-color: beige;
+  border: solid rgb(178, 178, 104);
+  border-radius: 10px;
+  display: flex;
+  justify-content: center;
+  margin: 1rem 24%;
+  align-items: center;
 }
 
 .tool {
-    border: none;
-    background-color: beige;
-    display: inline-block;
-    width: 2.4rem;
+  border: none;
+  background-color: beige;
+  display: inline-block;
+  width: 2.4rem;
 }
-.tool:nth-child(-n+4) {
-    width: 1.8rem;
+.tool:nth-child(-n + 4) {
+  width: 1.8rem;
 }
 
 .tool:hover {
-    border:  #ccc;
-    border-radius: 2px;
-    background-color: #fff;
+  border: #ccc;
+  border-radius: 2px;
+  background-color: #fff;
 }
 
 .divider {
-    display: inline-block;
-    text-indent: -9999px;
-    margin: 0 5px;
-    height: 100%;
-    border-right: 1.5px solid #ddd;
+  display: inline-block;
+  text-indent: -9999px;
+  margin: 0 5px;
+  height: 100%;
+  border-right: 1.5px solid #ddd;
 }
-
 
 /* 正文内容区 */
 .textRoot {
-    background-image: url("../image/暖色背景.jpg");
-    background-size: 100%;
-    padding-top: 1rem;
-    height: 100%;
+  background-image: url("../image/暖色背景.jpg");
+  background-size: 100%;
+  padding-top: 1rem;
+  height: 100%;
 }
 
 /*文本区*/
 .textArea {
-    /* padding-left: -15px;
+  /* padding-left: -15px;
     padding-right: -15px; */
-    overflow: hidden;
-    margin: 2rem 6rem;
-    height: 75%;
+  overflow: hidden;
+  margin: 2rem 6rem;
+  height: 75%;
 }
 
 /*文本输入区*/
 .textInput {
-    width: 50%;
-    float: left;
-    overflow: auto;
+  width: 50%;
+  float: left;
+  overflow: auto;
 }
 
 /*文本输入框*/
 .textarea {
-    padding: 1rem 1.3rem;
-    width: 100%;
-    height: 98%;
-    resize: none;
-    background-color: rgba(255, 255, 255, 0.7); /* 半透明白色背景 */
-    backdrop-filter: blur(5px);
-    -webkit-backdrop-filter: blur(10px);
-    border-radius: 15px;
-    border: solid 5px rgb(178, 178, 104);
-    word-break: break-all;
+  padding: 1rem 1.3rem;
+  width: 100%;
+  height: 98%;
+  resize: none;
+  background-color: rgba(255, 255, 255, 0.7); /* 半透明白色背景 */
+  backdrop-filter: blur(5px);
+  -webkit-backdrop-filter: blur(10px);
+  border-radius: 15px;
+  border: solid 5px rgb(178, 178, 104);
+  word-break: break-all;
 }
 
 textarea:focus {
-    outline: none;
-  }
-
+  outline: none;
+}
 
 /* 输出结果区？ */
 section {
-    height: 100%;
-    width: 50%;
-    padding-left: 15px;
-    overflow: auto;
-    float: right;
+  height: 100%;
+  width: 50%;
+  padding-left: 15px;
+  overflow: auto;
+  float: right;
 }
 
 .result-html {
-    overflow: auto;
+  overflow: auto;
 }
 
 /* 模态框 */
-.modal-dialog{
-    width: 100%;
+.modal-dialog {
+  width: 100%;
 }
 
-.my-modal{
-    background-color: aquamarine;
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%, -50%);
-    position: fixed;
-    border: rgb(178, 178, 104) solid 5px;
-    border-radius: 15px;
-    padding: 1rem 2rem;
+.my-modal {
+  background-color: aquamarine;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+  position: fixed;
+  border: rgb(178, 178, 104) solid 5px;
+  border-radius: 15px;
+  padding: 1rem 2rem;
 }
 
 .my-overlay {
-    /* 修改模态框遮罩层样式 */
-    position: fixed;
-    inset: 0px;
-    background-color: rgba(255, 255, 255, 0.3);
-  }
-
-.modal-button{
-    background-color:aquamarine;
-    border: rgb(178, 178, 104) solid 2px;
-    padding: .2rem;
-    border-radius: 5px;
-    float: right;
+  /* 修改模态框遮罩层样式 */
+  position: fixed;
+  inset: 0px;
+  background-color: rgba(255, 255, 255, 0.3);
 }
 
-.modal-header{
-    font-size: 20px;
-    font-weight: bold;
-    padding-bottom: .8rem;
+.modal-button {
+  background-color: aquamarine;
+  border: rgb(178, 178, 104) solid 2px;
+  padding: 0.2rem;
+  border-radius: 5px;
+  float: right;
 }
 
-
-.radios{
-    height: 35px;
-    padding-top: 10px;
-    margin-bottom: 10px;
-    /* display: flexbox; */
+.modal-header {
+  font-size: 20px;
+  font-weight: bold;
+  padding-bottom: 0.8rem;
 }
 
- /* label input[type="radio"] {
+.radios {
+  height: 35px;
+  padding-top: 10px;
+  margin-bottom: 10px;
+  /* display: flexbox; */
+}
+
+/* label input[type="radio"] {
     width: 20px; 
     height: 20px; 
   } */
-  
-  label i {
-    font-size: 24px; 
-  }
 
-  .modal-label{
-    width: 20%;
-    transform: scale(1.2);
-  }
-
-hr{
-    margin-top: 0;
-    margin-bottom: 1rem;
-    color: black;
+label i {
+  font-size: 24px;
 }
 
+.modal-label {
+  width: 20%;
+  transform: scale(1.2);
+}
+
+hr {
+  margin-top: 0;
+  margin-bottom: 1rem;
+  color: black;
+}
 
 .modal-table {
-    border-collapse: collapse; /* 合并单元格边框 */
-    width: 100%;
-  }
-  
-.modal-table th, .modal-table td {
-    border: 1px solid rgb(0, 0, 0); /* 设置单元格边框 */
-    padding: 3px 15px ;
-    text-align: left;
-  }
-  
+  border-collapse: collapse; /* 合并单元格边框 */
+  width: 100%;
+}
+
+.modal-table th,
+.modal-table td {
+  border: 1px solid rgb(0, 0, 0); /* 设置单元格边框 */
+  padding: 3px 15px;
+  text-align: left;
+}
