+++ src/main.tsx
@@ -5,10 +5,12 @@ import Modal from 'react-modal';
 
 
 const container = document.getElementById('root');
+
 if (container) {
   container.className = 'full-height';
 }
-if (container) { // 增加类型检查，确保 container 不为 null
+
+if (container) { 
   const root = createRoot(container);
   root.render(<App />);
 } else {
